Fix RBF gradient with scale_by_similarity
Previously it did not use error_vec /= self._total_similarity, as math
dictates for gradient, because it showed poor performance in empirical
tests. Turned out the issue was an error when returning error. The true
error and gradient were good, but the error displayed was inaccurate. This
has been fixed.

To better facilitate this fix, Model.train_step does not assume a vector
is returned by _train_increment, and perform mean squared error on it.
This also allows better control over error function.